refactor: place order_graph_validator in order directort
