+++ vda5050_core/include/vda5050_core/order/order_graph_validator.hpp
@@ -16,8 +16,8 @@
  * limitations under the License.
  */
 
-#ifndef VDA5050_CORE__ORDER_EXECUTION__ORDER_GRAPH_VALIDATOR_HPP_
-#define VDA5050_CORE__ORDER_EXECUTION__ORDER_GRAPH_VALIDATOR_HPP_
+#ifndef VDA5050_CORE__ORDER_GRAPH_VALIDATOR__ORDER_GRAPH_VALIDATOR_HPP_
+#define VDA5050_CORE__ORDER_GRAPH_VALIDATOR__ORDER_GRAPH_VALIDATOR_HPP_
 
 #include <vector>
 
@@ -62,4 +62,4 @@ private:
 
 }  // namespace order_graph_validator
 }  // namespace vda5050_core
-#endif  // VDA5050_CORE__ORDER_EXECUTION__ORDER_GRAPH_VALIDATOR_HPP_
+#endif  // VDA5050_CORE__ORDER_GRAPH_VALIDATOR__ORDER_GRAPH_VALIDATOR_HPP_
